Move vscode-uitls/external-files.ts to /common/vscode/external-files.ts
